// src/lib/enums/clock.enum.ts
export enum ClockEnum {
    SHIFT_START_DATE_AND_TIME = "shiftstartdateandtime",
    SHIFT_END_DATE_AND_TIME = "shiftenddateandtime",
    EXAMSTE_DOWNLOAD_TIME = "examstedownloadtime",
    QP_DELIVERY_TIME = "qpdeliverytime",
    USER_DATA_DELIVERY_TIME = "userdatadeliverytime",
    GATE_CLOSING_TIME = "gateclosingtime",
    LIVE_EXAM_START_TIME = "liveexamstarttime",
    BIOMETRIC_START_TIME = "biometricstarttime",
    CANDIDATE_START_TIME = "candidatestarttime",
    EXAM_START_TIME = "examstarttime",
    EXAM_END_TIME = "examendtime",
    MINIMUM_SUBMISSION_TIME = "minimumsubmissiontime",
    LOGIN_TIME = "logintime",
    // CANDIDATE_DATE_TIME = "candidate_date_time",
    SYNC_TIME = "synctime"

}